Add TestTelemetry
Running the following

```
go test ./telemetry
```

shows that the buffering logic in telemetry does not send all the events
to a telemetry endpoint.

This commit also no longer starts the telemetry goroutine on New() to allow
callers such as a test, to modify its fields before starting it.
Now invoking an explicit Start() method is needed after New().

Signed-off-by: Tibor Vass <teabee89@gmail.com>