mcp: implement a new pipe session attachable
So far we were using the terminal session attachable to pass data between
cli and engine for the MCP stdio server.

However, MCP clients when they execute the cli they don't provide a TTY.

The new pipe session attachable allows to send and receive data between
the cli and the engine.

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>
Signed-off-by: Tibor Vass <teabee89@gmail.com>